Move `createFakeController` function to test file  (#1148)
* Remove unused function


* Move createFakeController to test file


* Remove unused constant


---------
