+++ pkg/util/provider/drain/drain.go
@@ -132,10 +132,6 @@ const (
 	// DefaultMachineDrainTimeout is the default value for MachineDrainTimeout
 	DefaultMachineDrainTimeout = 2 * time.Hour
 
-	// PodsWithoutPVDrainGracePeriod defines the grace period to wait for the pods without PV during machine drain.
-	// This is in addition to the maximum terminationGracePeriod amount the pods.
-	PodsWithoutPVDrainGracePeriod = 3 * time.Minute
-
 	// Interval is the default Poll interval
 	Interval = time.Second * 5
 
@@ -473,18 +469,6 @@ func (o *Options) getTerminationGracePeriod(pod *corev1.Pod) time.Duration {
 	}
 
 	return time.Duration(*pod.Spec.TerminationGracePeriodSeconds) * time.Second
-}
-
-func (o *Options) getGlobalTimeoutForPodsWithoutPV(pods []*corev1.Pod) time.Duration {
-	var tgpsMax time.Duration
-	for _, pod := range pods {
-		tgps := o.getTerminationGracePeriod(pod)
-		if tgps > tgpsMax {
-			tgpsMax = tgps
-		}
-	}
-
-	return tgpsMax + PodsWithoutPVDrainGracePeriod
 }
 
 func (o *Options) evictPods(ctx context.Context, attemptEvict bool, pods []corev1.Pod, policyGroupVersion string, getPodFn func(namespace, name string) (*corev1.Pod, error)) error {
